Combine layouts into one layout
This commit removes BaseLayout, AnonymousLayout and Dashboard

BaseLayout and Dashboard are merged with MainLayout
Breadcrumbs are hidden for anonymous routes now under MainLayout
