--- frontend/src/views/Dashboard.vue
@@ -1,69 +1,0 @@
-/* * sri sri guru gauranga jayatah */
-
-<template>
-  <main-layout>
-    <template slot="aboveRoute">
-      <v-breadcrumbs :items="breadcrumbs" divider=">"></v-breadcrumbs>
-    </template>
-  </main-layout>
-</template>
-
-<script lang="ts">
-import { Component, Vue, Watch } from "vue-property-decorator";
-import MainLayout from "@/views/Layout/MainLayout.vue";
-import _ from "lodash";
-
-const pathMap = {
-  cr: "CR",
-  sqr: "SQR",
-  te: "TE",
-  allot: "Allotment"
-};
-
-@Component({
-  name: "Dashboard",
-  components: {
-    MainLayout
-  }
-})
-export default class Dashboard extends Vue {
-  breadcrumbs: any[] = [];
-
-  @Watch("$route", { immediate: true, deep: true })
-  getBreadcrumbs({ path, params }: any) {
-    const breadcrumbs: any[] = [];
-    if (path !== "/") {
-      const paths = path.split("/");
-      const pathsLength = paths.length - 1;
-      paths.forEach((item: string, index: number) => {
-        breadcrumbs.push({
-          text: this.getText(item, params),
-          disabled: index === pathsLength,
-          to: this.getPath(item, index, paths),
-          exact: true
-        });
-      });
-    }
-    this.breadcrumbs = breadcrumbs;
-  }
-
-  getPath(item: string, index: number, paths: string[]) {
-    return `${paths.slice(0, index).join("/")}/${item}`;
-  }
-
-  getText(path: string, params: any) {
-    if (path) {
-      let customText = _.get(
-        pathMap,
-        path,
-        path[0].toUpperCase() + path.substring(1)
-      );
-      if (typeof customText === "function") {
-        customText = customText(params);
-      }
-      return customText;
-    }
-    return "Home";
-  }
-}
-</script>
