+++ frontend/src/views/Layout/MainLayout.vue
@@ -1,95 +1,105 @@
 <template>
-  <base-layout>
-    <template slot="aboveContent">
-      <v-navigation-drawer
-        fixed
-        :clipped="$vuetify.breakpoint.lgAndUp"
-        v-model="sidebar"
-        app
-        v-if="currentUser"
-        width="300px"
-      >
-        <v-list class="pa-0">
-          <v-list-item>
-            <v-list-item-avatar>
-              <img :src="currentUser.photoURL" />
-            </v-list-item-avatar>
+  <v-app>
+    <v-navigation-drawer
+      fixed
+      :clipped="$vuetify.breakpoint.lgAndUp"
+      v-model="sidebar"
+      app
+      v-if="currentUser"
+      width="300px"
+    >
+      <v-list class="pa-0">
+        <v-list-item>
+          <v-list-item-avatar>
+            <img :src="currentUser.photoURL" />
+          </v-list-item-avatar>
+          <v-list-item-content>
+            <v-list-item-title>{{ currentUser.displayName }}</v-list-item-title>
+          </v-list-item-content>
+          <v-list-item-action>
+            <v-btn color="rgba(0, 0, 0, 0.87)" icon ripple @click="signOut">
+              <v-icon>fas fa-sign-out-alt</v-icon>
+            </v-btn>
+          </v-list-item-action>
+        </v-list-item>
+      </v-list>
+      <v-list dense expand>
+        <v-divider></v-divider>
+        <template v-for="(item, index) in menuItems">
+          <menu-links
+            :parentRoute="item"
+            :key="index"
+            v-if="item.meta && item.meta.activator"
+            :activeClass="getActiveClass(item.path)"
+          >
+          </menu-links>
+          <v-list-item
+            :to="`/${item.path}`"
+            v-else-if="item.meta.menuItem"
+            :key="`no-nested-${index}`"
+            :active-class="getActiveClass(item.path)"
+          >
+            <v-list-item-icon>
+              <v-icon v-text="item.meta.menuIcon"></v-icon>
+            </v-list-item-icon>
             <v-list-item-content>
-              <v-list-item-title>{{
-                currentUser.displayName
-              }}</v-list-item-title>
+              <v-list-item-title>{{ item.meta.menuName }}</v-list-item-title>
             </v-list-item-content>
-            <v-list-item-action>
-              <v-btn color="rgba(0, 0, 0, 0.87)" icon ripple @click="signOut">
-                <v-icon>fas fa-sign-out-alt</v-icon>
-              </v-btn>
-            </v-list-item-action>
           </v-list-item>
-        </v-list>
-        <v-list dense expand>
-          <v-divider></v-divider>
-          <template v-for="(item, index) in menuItems">
-            <menu-links
-              :parentRoute="item"
-              :key="index"
-              v-if="item.meta && item.meta.activator"
-              :activeClass="getActiveClass(item.path)"
-            >
-            </menu-links>
-            <v-list-item
-              :to="`/${item.path}`"
-              v-else-if="item.meta.menuItem"
-              :key="`no-nested-${index}`"
-              :active-class="getActiveClass(item.path)"
-            >
-              <v-list-item-icon>
-                <v-icon v-text="item.meta.menuIcon"></v-icon>
-              </v-list-item-icon>
-              <v-list-item-content>
-                <v-list-item-title>{{ item.meta.menuName }}</v-list-item-title>
-              </v-list-item-content>
-            </v-list-item>
-          </template>
-          <v-divider></v-divider>
-          <v-list-item @click="signOut" v-if="currentUser">
-            <v-list-item-action>
-              <v-icon left>fas fa-sign-out-alt</v-icon>
-            </v-list-item-action>
-            <v-list-item-content>SignOut</v-list-item-content>
-          </v-list-item>
-        </v-list>
-      </v-navigation-drawer>
+        </template>
+        <v-divider></v-divider>
+        <v-list-item @click="signOut" v-if="currentUser">
+          <v-list-item-action>
+            <v-icon left>fas fa-sign-out-alt</v-icon>
+          </v-list-item-action>
+          <v-list-item-content>SignOut</v-list-item-content>
+        </v-list-item>
+      </v-list>
+    </v-navigation-drawer>
 
-      <v-app-bar :clipped-left="$vuetify.breakpoint.lgAndUp" fixed app>
-        <v-app-bar-nav-icon
-          @click="sidebar = !sidebar"
-          v-if="currentUser"
-        ></v-app-bar-nav-icon>
-        <v-toolbar-title>
-          <router-link to="/" tag="span" style="cursor: pointer">{{
-            appTitle
-          }}</router-link>
-        </v-toolbar-title>
-      </v-app-bar>
-    </template>
-    <template slot="aboveRoute">
-      <slot name="aboveRoute"></slot>
-    </template>
-  </base-layout>
+    <v-app-bar :clipped-left="$vuetify.breakpoint.lgAndUp" fixed app>
+      <v-app-bar-nav-icon
+        @click="sidebar = !sidebar"
+        v-if="currentUser"
+      ></v-app-bar-nav-icon>
+      <v-toolbar-title>
+        <router-link to="/" tag="span" style="cursor: pointer">{{
+          appTitle
+        }}</router-link>
+      </v-toolbar-title>
+    </v-app-bar>
+    <v-content>
+      <v-container fluid>
+        <v-breadcrumbs
+          v-if="breadcrumbs.length"
+          :items="breadcrumbs"
+          divider=">"
+        ></v-breadcrumbs>
+        <router-view></router-view>
+      </v-container>
+    </v-content>
+  </v-app>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
+import { Component, Vue, Watch } from "vue-property-decorator";
+import _ from "lodash";
 import { mapState, mapActions } from "vuex";
 import { RouteConfig } from "vue-router";
+import firebase from "firebase/app";
 import { getMenuItems } from "@/router";
-import BaseLayout from "./BaseLayout.vue";
 import MenuLinks from "@/components/MenuLinks";
+
+const pathMap = {
+  cr: "CR",
+  sqr: "SQR",
+  te: "TE",
+  allot: "Allotment"
+};
 
 @Component({
   name: "MainLayout",
   components: {
-    BaseLayout,
     MenuLinks
   },
   computed: {
@@ -103,9 +113,63 @@ export default class MainLayout extends Vue {
   appTitle = "Audio Seva";
   sidebar = false;
   menuItems: RouteConfig[] = [];
+  breadcrumbs: any[] = [];
 
-  async mounted() {
-    this.menuItems = await getMenuItems();
+  @Watch("currentUser", { immediate: true })
+  async handleCurrentUserChange(newUser: firebase.User | null) {
+    if (newUser) {
+      this.menuItems = await getMenuItems();
+    } else {
+      this.menuItems = [];
+    }
+  }
+
+  @Watch("$route", { immediate: true, deep: true })
+  getBreadcrumbs({ path, params, meta }: any) {
+    const breadcrumbs: any[] = [];
+    /*
+     * Disable breadcrumbs for anonymous routes
+     * Check explicitly for false because children of authenticated routes might not contain information about authentication
+     */
+    if (
+      _.get(meta, "auth.requireAuth") === false ||
+      _.get(meta, "auth.guestOnly")
+    ) {
+      this.breadcrumbs = breadcrumbs;
+      return;
+    }
+    if (path !== "/") {
+      const paths = path.split("/");
+      const pathsLength = paths.length - 1;
+      paths.forEach((item: string, index: number) => {
+        breadcrumbs.push({
+          text: this.getText(item, params),
+          disabled: index === pathsLength,
+          to: this.getPath(item, index, paths),
+          exact: true
+        });
+      });
+    }
+    this.breadcrumbs = breadcrumbs;
+  }
+
+  getPath(item: string, index: number, paths: string[]) {
+    return `${paths.slice(0, index).join("/")}/${item}`;
+  }
+
+  getText(path: string, params: any) {
+    if (path) {
+      let customText = _.get(
+        pathMap,
+        path,
+        path[0].toUpperCase() + path.substring(1)
+      );
+      if (typeof customText === "function") {
+        customText = customText(params);
+      }
+      return customText;
+    }
+    return "Home";
   }
 
   getActiveClass(path: string) {
@@ -119,5 +183,9 @@ export default class MainLayout extends Vue {
 <style scoped>
 >>> .inactive-menu {
   color: rgba(0, 0, 0, 0.87);
+}
+
+>>> .v-content__wrap {
+  display: flex;
 }
 </style>
